requirements.txt
fastapi[standard]==0.115.0
faster-whisper==1.1.0
python-multipart==0.0.9
pydantic==2.8.0
pydantic-settings==2.4.0
uvicorn[standard]==0.32.0
websockets==12.0
webrtcvad-wheels==2.0.10.post1
redis==5.0.7
rq==1.16.2
ffmpeg-python==0.2.0
python-dotenv==1.0.1

# NVIDIA libraries for CUDA support
nvidia-cublas-cu12
nvidia-cudnn-cu12==9.*

# Development and testing dependencies
pytest==8.0.0
httpx==0.27.0
